+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved `Step.ensure_result()` such that the step's result doesn't have to be read from the cache.
 - Fixed an issue with the output from `MulticoreExecutor` such that it's now consistent with the default `Executor` for steps that were found in the cache.
 - One of our error messages referred to a configuration file that no longer exists.
+- Improved performance of `BeakerWorkspace`.
 
 
 ## [v0.9.1](https://github.com/allenai/tango/releases/tag/v0.9.1) - 2022-06-24